[improvement] Showed verses of a whole chapter if verses from another chapter are also included in a reference (10m)
This was an issue where the well-formatted reference ended up being "Genesis 2:8-9; 3", which when parsed back into Pericope ended up as "Genesis 2:8-9, 3" (verse 3 instead of chapter 3). It seemed more fraught to try to scale back some of the generosity of the parsing engine to parse the reference correctly than it would be to tweak the output rules to clarify the entirety of the chapter, which seems to be the way references would want to be displayed in such cases anyhow.